cephadm: move basic exception types to exceptions.py
Pair-programmed-with: Adam King <adking@redhat.com>
